Fix newlines and javadoc
