+++ src/main/java/im/tox/upsourcebot/client/tasks/RecoveringCallable.java
@@ -3,11 +3,17 @@ package im.tox.upsourcebot.client.tasks;
 /**
  * Callable with a specific exception to recover from.
  *
- * @param <E> Exception type recoverable by this Callable. If this exception is thrown, the callable
- *            can be retried at a later time.
+ * @param <E> Exception type recoverable by this Callable.
  */
 public interface RecoveringCallable<E extends Exception> {
 
+  /**
+   * Does any computation
+   *
+   * @throws E                              if the computation failed. This exception is
+   *                                        recoverable, and this callable can be retried.
+   * @throws java.lang.InterruptedException if the thread is interrupted.
+   */
   void call() throws E, InterruptedException;
 
 }
